+++ common/Animations/Keyframe.cs
@@ -5,32 +5,32 @@ namespace StorybrewCommon.Animations;
 
 /// <summary> Represents a keyframe with a time and value. </summary>
 /// <typeparam name="TValue"> A type that represents the value of this keyframe. </typeparam>
-public readonly struct Keyframe<TValue>(double time, TValue value, Func<double, double> easing, bool until) : IEquatable<Keyframe<TValue>>, IComparer<Keyframe<TValue>>
+public readonly struct Keyframe<TValue>(float time, TValue value, Func<float, float> easing, bool until) : IEquatable<Keyframe<TValue>>, IComparer<Keyframe<TValue>>
 {
     ///<summary> Time of this keyframe. </summary>
-    public readonly double Time = time;
+    public readonly float Time = time;
 
     ///<summary> Value of this keyframe. </summary>
     public readonly TValue Value = value;
 
     ///<summary> <see cref="EasingFunctions"/> easing of this keyframe. </summary>
-    public readonly Func<double, double> Ease = easing ?? EasingFunctions.Linear;
+    public readonly Func<float, float> Ease = easing ?? EasingFunctions.Linear;
 
     ///<summary> Reserved for <see cref="Storyboarding.Util.CommandGenerator"/>. </summary>
     internal readonly bool Until = until;
 
     ///<summary> Initializes a new keyframe with a default value. </summary>
     ///<param name="time"> Time of the keyframe. </param>
-    public Keyframe(double time) : this(time, default) { }
+    public Keyframe(float time) : this(time, default) { }
 
     ///<summary> Initializes a new keyframe. </summary>
     ///<param name="time"> Time of the keyframe. </param>
     ///<param name="value"> A value to be assigned to the keyframe. </param>
     ///<param name="easing"> <see cref="EasingFunctions"/> easing to be assigned. </param>
-    public Keyframe(double time, TValue value, Func<double, double> easing = null) : this(time, value, easing, false) { }
+    public Keyframe(float time, TValue value, Func<float, float> easing = null) : this(time, value, easing, false) { }
 
     ///<summary> Overrides a keyframe's time. </summary>
-    public Keyframe<TValue> WithTime(double time) => new(time, Value, Ease);
+    public Keyframe<TValue> WithTime(float time) => new(time, Value, Ease);
 
     ///<summary> Overrides a keyframe's value. </summary>
     public Keyframe<TValue> WithValue(TValue value) => new(Time, value, Ease);
