+++ common/Scripting/GeneratorContext.cs
@@ -18,8 +18,8 @@ public abstract class GeneratorContext
     public abstract IEnumerable<Beatmap> Beatmaps { get; }
     public abstract StoryboardLayer GetLayer(string identifier);
 
-    public abstract double AudioDuration { get; }
-    public abstract float[] GetFft(double time, string path = null, bool splitChannels = false);
+    public abstract float AudioDuration { get; }
+    public abstract float[] GetFft(float time, string path = null, bool splitChannels = false);
     public abstract float GetFftFrequency(string path = null);
 
     public abstract bool Multithreaded { get; set; }
