+++ common/Storyboarding/Commands/ColorCommand.cs
@@ -3,11 +3,11 @@
 namespace StorybrewCommon.Storyboarding.Commands;
 
 #pragma warning disable CS1591
-public class ColorCommand(OsbEasing easing, double startTime, double endTime, CommandColor startValue, CommandColor endValue) : Command<CommandColor>("C", easing, startTime, endTime, startValue, endValue)
+public class ColorCommand(OsbEasing easing, float startTime, float endTime, CommandColor startValue, CommandColor endValue) : Command<CommandColor>("C", easing, startTime, endTime, startValue, endValue)
 {
-    public override CommandColor ValueAtProgress(double progress) => StartValue + (EndValue - StartValue) * progress;
-    public override CommandColor Midpoint(Command<CommandColor> endCommand, double progress) => StartValue + (endCommand.EndValue - StartValue) * progress;
-    public override IFragmentableCommand GetFragment(double startTime, double endTime)
+    public override CommandColor ValueAtProgress(float progress) => StartValue + (EndValue - StartValue) * progress;
+    public override CommandColor Midpoint(Command<CommandColor> endCommand, float progress) => StartValue + (endCommand.EndValue - StartValue) * progress;
+    public override IFragmentableCommand GetFragment(float startTime, float endTime)
     {
         if (IsFragmentable)
         {
