+++ common/Storyboarding/Commands/ITypedCommand.cs
@@ -6,7 +6,7 @@ public interface ITypedCommand<TValue> : ICommand
     OsbEasing Easing { get; }
     TValue StartValue { get; }
     TValue EndValue { get; }
-    double Duration { get; }
+    float Duration { get; }
 
-    TValue ValueAtTime(double time);
+    TValue ValueAtTime(float time);
 }
